# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.0
paths:
  /tasks:
    get:
      summary: Get all tasks for a specific user
      tags:
        - tasks
      parameters:
        - name: user_id
          in: query
          description: The ID of the user
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: A list of tasks for the specified user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create a new task
      tags:
        - tasks
      requestBody:
        description: The task to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '201':
          description: The created task ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
    patch:
      summary: Update an existing task
      tags:
        - tasks
      parameters:
        - name: id
          in: query
          description: The ID of the task
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: The task to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '200':
          description: The updated task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    delete:
      summary: Delete an existing task by Id
      tags:
        - tasks
      parameters:
        - name: id
          in: query
          description: The ID of the task
          required: true
          schema:
            type: integer
            format: int64
        - name: user_id
          in: query
          description: The ID of the user
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: The task was successfully deleted

  /users:
    post:
      summary: Create a new user
      tags:
        - users
      requestBody:
        description: The user to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: The created user ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
    patch:
      summary: Update an existing user
      tags:
        - users
      parameters:
        - name: id
          in: query
          description: The ID of the user
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: The user to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: The updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      summary: Delete an existing user by Id
      tags:
        - users
      parameters:
        - name: id
          in: query
          description: The ID of the user
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: The user was successfully deleted

components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        title:
          type: string
        is_done:
          type: boolean

    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        password:
          type: string

    IdResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64